add now playing button to album tracklist screen
