+++ lib/pages/album_tracklist.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:music_home/widgets/now_playing_button.dart';
 import '../data/song_data.dart';
 import '../widgets/mp_listview.dart';
 
@@ -13,6 +14,9 @@ class AlbumTrackList extends StatelessWidget {
     return Scaffold(
       appBar: AppBar(
         title: Text(album.title),
+        actions: <Widget>[
+          NowPlayingButton(),
+        ],
       ),
       body: MPListView(data, false, tracks: album.tracks),
     );
