+++ lib/pages/root_page.dart
@@ -1,31 +1,18 @@
 import 'package:flutter/material.dart';
-import 'package:music_home/data/song_data.dart';
+import 'package:music_home/widgets/now_playing_button.dart';
 
 import '../widgets/mp_inherited.dart';
 import '../widgets/mp_listview.dart';
-import 'now_playing.dart';
 
 class RootPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final rootIW = MPInheritedWidget.of(context);
-    //Goto Now Playing Page
-    void goToNowPlaying(Track s, {bool nowPlayTap: false}) {
-      Navigator.push(
-        context,
-        MaterialPageRoute<void>(
-          builder: (context) => NowPlaying(
-            rootIW.songData,
-            s,
-            nowPlayTap: nowPlayTap,
-          ),
-        ),
-      );
-    }
 
     //Shuffle Songs and goto now playing page
     void shuffleSongs() {
-      goToNowPlaying(rootIW.songData.randomSong);
+      // FIXME: need to move this into songdata
+      //goToNowPlaying(rootIW.songData.randomSong);
     }
 
     return DefaultTabController(
@@ -34,17 +21,7 @@ class RootPage extends StatelessWidget {
         appBar: AppBar(
           title: Text('Flutter Music Player'),
           actions: <Widget>[
-            Container(
-              padding: const EdgeInsets.all(20.0),
-              child: Center(
-                child: InkWell(
-                    child: Text('Now Playing'),
-                    onTap: () => goToNowPlaying(
-                          rootIW.songData.currentlyPlayingTrack,
-                          nowPlayTap: true,
-                        )),
-              ),
-            )
+            NowPlayingButton(),
           ],
           bottom: TabBar(tabs: [
             Tab(text: 'All Songs'),
